Move _ThrottledFetcher tests to tests/ to fix Qt test isolation
tuf.ngclient.RequestsFetcher.__init__ (called via _ThrottledFetcher
super().__init__) corrupts state that QSystemTrayIcon needs on Python
3.14 + PySide6 6.11 when run in the same process before the tray is
constructed. _ThrottledFetcher tests need no Qt so they belong in
tests/ not tests-qt/.
